requirements.txt
alabaster==0.7.12
amqp==2.6.0
appdirs==1.4.4
argon2-cffi==19.2.0
aspy.yaml==1.3.0
astroid==2.4.2
atomicwrites==1.4.0
attrs==19.3.0
Babel==2.8.0
backcall==0.2.0
beautifulsoup4==4.9.1
billiard==3.6.3.0
black==19.10b0
cached-property==1.5.1
celery==4.3.0
certifi==2020.6.20
cffi==1.14.0
cfgv==3.1.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.4
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==2.2.7
django-allauth==0.40.0
django-annoying==0.10.6
django-bootstrap4==2.1.1
django-celery-beat==1.5.0
django-ckeditor==5.9.0
django-coverage-plugin==1.6.0
django-crispy-forms==1.8.1
django-debug-toolbar==2.1
django-environ==0.4.5
django-extensions==2.2.5
django-extra-views==0.13.0
django-filter==2.3.0
django-js-asset==1.2.2
django-jsonstore==0.4.1
django-material==1.6.7
django-model-utils==3.2.0
django-ranged-response==0.2.0
django-redis==4.10.0
django-simple-captcha==0.5.12
django-timezone-field==4.0
django-viewflow==1.6.0
djangorestframework==3.10.3
docutils==0.16
entrypoints==0.3
factory-boy==2.12.0
Faker==4.1.1
filelock==3.0.12
flake8==3.7.8
flower==0.9.3
identify==1.4.20
idna==2.10
imagesize==1.2.0
ipdb==0.12.2
ipython==7.16.1
ipython-genutils==0.2.0
isodate==0.6.0
isort==4.3.21
itypes==1.2.0
jedi==0.17.1
Jinja2==2.11.2
kombu==4.6.11
lazy-object-proxy==1.4.3
lxml==4.5.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.4.0
mypy==0.740
mypy-extensions==0.4.3
nodeenv==1.4.0
oauthlib==3.1.0
packaging==20.4
parso==0.7.0
pathspec==0.8.0
pickleshare==0.7.5
Pillow==6.2.1
pluggy==0.13.1
pre-commit==1.20.0
prompt-toolkit==3.0.5
psycopg2==2.8.5
psycopg2-binary==2.8.5
py==1.9.0
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
Pygments==2.6.1
pylint==2.5.3
pylint-celery==0.3
pylint-django==2.0.13
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==5.3.0
pytest-django==3.7.0
pytest-sugar==0.9.2
python-crontab==2.5.1
python-dateutil==2.8.1
python-slugify==4.0.0
python3-openid==3.2.0
pytz==2019.3
PyYAML==5.3.1
redis==3.3.11
regex==2020.6.8
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.15.0
snowballstemmer==2.0.0
soupsieve==2.0.1
Sphinx==2.2.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.1
tornado==5.1.1
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.2
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
virtualenv==20.0.25
wcwidth==0.2.5
Werkzeug==0.14.1
whitenoise==4.1.4
wrapt==1.12.1
xmltodict==0.12.0
zeep==3.4.0
